backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
mysql-connector-python==8.1.0
bcrypt==4.0.1
qrcode==7.4.2
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7